requirements.txt
discord.py>=2.0.0
requests>=2.25.1
python-dotenv>=0.20.0